excel reader: parse account number

+++ pkg/migrations/sql/20250515175335_initial.sql
@@ -9,7 +9,7 @@ CREATE TABLE rules (
 CREATE TABLE transactions (
     id SERIAL PRIMARY KEY,
     hash VARCHAR(100) NOT NULL,
-    account SMALLINT NOT NULL,
+    account BIGSERIAL NOT NULL,
     date TIMESTAMP NOT NULL,
     kind VARCHAR(30) NOT NULL,
     content TEXT NOT NULL,
